editing tests and services

+++ scrapper/src/main/java/edu/java/service/chat/jdbc/JdbcChatService.java
@@ -1,29 +1,33 @@
 package edu.java.service.chat.jdbc;
 
+import edu.java.exceptions.ChatAlreadyRegisteredException;
+import edu.java.exceptions.ChatNotFoundException;
 import edu.java.repository.chat.ChatRepository;
 import edu.java.service.chat.ChatService;
 import lombok.RequiredArgsConstructor;
-import lombok.extern.log4j.Log4j2;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
 @Service
 @RequiredArgsConstructor
-@Log4j2
 public class JdbcChatService implements ChatService {
     private final ChatRepository chatRepository;
 
     @Override
     @Transactional
     public void registerChat(Long chatId) {
-        log.info("Chat was registered successfully!");
+        if (chatRepository.isInTable(chatId)) {
+            throw new ChatAlreadyRegisteredException();
+        }
         chatRepository.add(chatId);
     }
 
     @Override
     @Transactional
     public void deleteChat(Long chatId) {
-        log.info("Chat was deleted successfully!");
+        if (!chatRepository.isInTable(chatId)) {
+            throw new ChatNotFoundException(chatId);
+        }
         chatRepository.remove(chatId);
     }
 }
